Major changes made to the following: Threading, Camera, Models, Input/movement.
Changed model system to a mesh system with a self made format .mesh , this means i can load in models of big sizes and not worry about loading and readin in a slow .txt file. I also added simplistic threading to the render frame and input handling. this means that the cpu has more space for working on all of the other processes instead of worrying about input andling and calculating. this now happens on a seperate thread. also took the draw call out of the shaders and put them into the mesh class, this makes creating new objects a lot cleaner, example: setting the shader to one type and drawing all following meshes, before i had to call the shader every time i wanted to draw a new mesh. finally upgraded the camera system to follow around my cursor, this uses yaw and pitch, meaning it can get stuck due to euler angles, an improvement would be quaternions, but that is for another time.